requirements.txt
Flask==2.2.2
torch==1.13.1
opencv-python==4.7.0.72
Pillow==9.3.0
